pyproject.toml
[project]
name = "farmtech-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "awscli>=1.40.18",
    "blue>=0.9.1",
    "boto3>=1.38.19",
    "fastapi[standard]>=0.115.12",
    "isort>=6.0.1",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "python-decouple>=3.8",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy-utils>=0.41.2",
    "streamlit>=1.45.0",
    "uvicorn>=0.34.2",
]
